oas-api/requirements.txt
bip-utils==1.7.0
Flask~=1.1
Flask-Cors==3.0.10
Flask-JSONRPC~=1.1.0
Flask-Uploads
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth==1.27.0
gunicorn==20.0.4
hdwallet~=0.2.0
ipfshttpclient==0.8.0a2
python-dotenv==0.19.1
Pillow~=8.1.0
pip~=21.0.1
py_crypto_hd_wallet
pyodbc~=4.0.30
PyMySQL==1.0.2
redis==3.2.1
requests
simplejson==3.17.2
SQLAlchemy==1.3.23
web3~=5.24.0
#only need this for IIS development
#steps here https://pypi.org/project/wfastcgi/
#wfastcgi~=3.0.0
#if required for additional firebase features
#firebase-admin
